+++ app/public/login.html
@@ -77,7 +77,7 @@
             };
     
             try {
-                const response = await fetch('/login', {
+                const response = await fetch('/account/login', {
                     method: 'POST',
                     headers: {
                         'Content-Type': 'application/json'
