+++ app/public/signup.html
@@ -90,7 +90,7 @@
             };
     
             try {
-                const response = await fetch('/signup', {
+                const response = await fetch('/account/register', {
                     method: 'POST',
                     headers: {
                         'Content-Type': 'application/json'
